Integrates register-DataAsset functionality into full socrata update pipeline and adds datasource_name to existing update DAGS.
