+++ airflow/dags/cook_county/update_chicago_relocated_vehicles.py
@@ -19,6 +19,7 @@ def update_chicago_relocated_vehicles():
     update_1 = update_socrata_table(
         socrata_table=SOCRATA_TABLE,
         conn_id="dwh_db_conn",
+        datasource_name="fluent_dwh_source",
         task_logger=task_logger,
     )
     update_1
