+++ airflow/dags/cook_county/update_cook_county_parcel_assessment_appeals.py
@@ -19,6 +19,7 @@ def update_cook_county_parcel_assessment_appeals():
     update_1 = update_socrata_table(
         socrata_table=SOCRATA_TABLE,
         conn_id="dwh_db_conn",
+        datasource_name="fluent_dwh_source",
         task_logger=task_logger,
     )
     update_1
