+++ airflow/great_expectations/great_expectations.yml
@@ -16,30 +16,8 @@ config_version: 3.0
 datasources:
   where_house_source:
     class_name: Datasource
-    data_connectors:
-      default_runtime_data_connector_name:
-        class_name: RuntimeDataConnector
-        name: default_runtime_data_connector_name
-        batch_identifiers:
-          - default_identifier_name
-        module_name: great_expectations.datasource.data_connector
-      data_raw_inferred_data_connector_name:
-        class_name: InferredAssetSqlDataConnector
-        name: data_raw_inferred_data_connector_name
-        module_name: great_expectations.datasource.data_connector
-        include_schema_name: true
-        introspection_directives:
-          schema_name: data_raw
-      dwh_inferred_data_connector_name:
-        class_name: InferredAssetSqlDataConnector
-        name: dwh_inferred_data_connector_name
-        module_name: great_expectations.datasource.data_connector
-        include_schema_name: true
-        introspection_directives:
-          schema_name: dwh
     execution_engine:
       module_name: great_expectations.execution_engine
-      class_name: SqlAlchemyExecutionEngine
       credentials:
         host: dwh_db
         port: '5432'
@@ -47,7 +25,29 @@ datasources:
         password: ${DWH_POSTGRES_PASSWORD}
         database: ${DWH_POSTGRES_DB}
         drivername: postgresql
+      class_name: SqlAlchemyExecutionEngine
     module_name: great_expectations.datasource
+    data_connectors:
+      default_runtime_data_connector_name:
+        class_name: RuntimeDataConnector
+        module_name: great_expectations.datasource.data_connector
+        name: default_runtime_data_connector_name
+        batch_identifiers:
+          - default_identifier_name
+      data_raw_inferred_data_connector_name:
+        class_name: InferredAssetSqlDataConnector
+        include_schema_name: true
+        module_name: great_expectations.datasource.data_connector
+        name: data_raw_inferred_data_connector_name
+        introspection_directives:
+          schema_name: data_raw
+      dwh_inferred_data_connector_name:
+        class_name: InferredAssetSqlDataConnector
+        include_schema_name: true
+        module_name: great_expectations.datasource.data_connector
+        name: dwh_inferred_data_connector_name
+        introspection_directives:
+          schema_name: dwh
 config_variables_file_path: uncommitted/config_variables.yml
 
 # The plugins_directory will be added to your python path for custom modules
@@ -113,8 +113,8 @@ anonymous_usage_statistics:
   data_context_id: b25cb2e2-4999-427c-93b4-7597f9e172a9
   enabled: true
 include_rendered_content:
-  expectation_suite: false
   globally: false
+  expectation_suite: false
   expectation_validation_result: false
 notebooks:
 fluent_datasources:
@@ -138,5 +138,17 @@ fluent_datasources:
         order_by: []
         batch_metadata: {}
         table_name: cook_county_sao_case_sentencing_data
+        schema_name: data_raw
+      data_raw.temp_cook_county_sao_case_disposition_data:
+        type: table
+        order_by: []
+        batch_metadata: {}
+        table_name: temp_cook_county_sao_case_disposition_data
+        schema_name: data_raw
+      data_raw.cook_county_sao_case_disposition_data:
+        type: table
+        order_by: []
+        batch_metadata: {}
+        table_name: cook_county_sao_case_disposition_data
         schema_name: data_raw
     connection_string: ${GX_DWH_DB_CONN}
